+++ game_patch/rf/misc.h
@@ -9,5 +9,5 @@ namespace rf
 
     static auto& default_player_weapon = addr_as_ref<String>(0x007C7600);
 
-    static auto& get_file_checksum = addr_as_ref<int(const char* filename)>(0x00436630);
+    static auto& get_file_checksum = addr_as_ref<unsigned(const char* filename)>(0x00436630);
 }